# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Campaign Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CampaignInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @minLength 1 */
  companyName: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @minLength 10 */
  productDescription: string;
  /** @minLength 5 */
  targetAudience: string;
  /** @minimum 1 */
  budget: number;
}

export interface ChannelAllocation {
  channel: string;
  percentage: number;
  amount: number;
}

export interface ABTestVariant {
  angle: string;
  type: string;
  hypothesis: string;
  primaryMetric: string;
}

export interface ABTestPlanItem {
  channel: string;
  variantA: ABTestVariant;
  variantB: ABTestVariant;
}

export interface CampaignIdea {
  title: string;
  description: string;
  rationale: string;
}

export interface AdCopyItem {
  channel: string;
  headline: string;
  body: string;
  callToAction: string;
}

export interface CampaignResult {
  channelMix: string[];
  budgetAllocation: ChannelAllocation[];
  abTestPlan: ABTestPlanItem[];
  /** @nullable */
  campaignIdeas?: CampaignIdea[] | null;
  /** @nullable */
  adCopy?: AdCopyItem[] | null;
  fitScore: number;
  /**
     * Deterministic risk assessment: low / medium / high
     * @nullable
     */
  riskLevel?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export type CampaignDetailStatus = typeof CampaignDetailStatus[keyof typeof CampaignDetailStatus];


export const CampaignDetailStatus = {
  processing: 'processing',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface CampaignDetail {
  id: number;
  status: CampaignDetailStatus;
  budget: number;
  productDescription?: string;
  targetAudience?: string;
  createdAt: string;
  result?: CampaignResult;
}

export type CampaignSubmitResponseStatus = typeof CampaignSubmitResponseStatus[keyof typeof CampaignSubmitResponseStatus];


export const CampaignSubmitResponseStatus = {
  processing: 'processing',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface CampaignSubmitResponse {
  campaignId: number;
  status: CampaignSubmitResponseStatus;
  result?: CampaignResult;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminAuthResponse {
  authenticated: boolean;
}

export type AdminCampaignListItemStatus = typeof AdminCampaignListItemStatus[keyof typeof AdminCampaignListItemStatus];


export const AdminCampaignListItemStatus = {
  processing: 'processing',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface AdminCampaignListItem {
  id: number;
  leadName: string;
  companyName: string;
  budget: number;
  status: AdminCampaignListItemStatus;
  /** @nullable */
  fitScore?: number | null;
  /**
     * Deterministic risk assessment: low / medium / high
     * @nullable
     */
  riskLevel?: string | null;
  createdAt: string;
}

export interface AdminCampaignListResponse {
  campaigns: AdminCampaignListItem[];
  total: number;
  page?: number;
  limit?: number;
}

export type AdminCampaignDetailStatus = typeof AdminCampaignDetailStatus[keyof typeof AdminCampaignDetailStatus];


export const AdminCampaignDetailStatus = {
  processing: 'processing',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface AdminCampaignDetail {
  id: number;
  leadName: string;
  leadEmail: string;
  companyName: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  phone?: string | null;
  budget: number;
  productDescription: string;
  targetAudience: string;
  status: AdminCampaignDetailStatus;
  createdAt: string;
  result?: CampaignResult;
  /** @nullable */
  rawAiResponse?: string | null;
}

export type AdminStatsTopChannelsItem = {
  channel: string;
  count: number;
};

export type AdminStatsStatusBreakdown = {
  processing: number;
  complete: number;
  failed: number;
};

export interface AdminStats {
  totalSubmissions: number;
  completionRate: number;
  averageBudget: number;
  topChannels: AdminStatsTopChannelsItem[];
  statusBreakdown: AdminStatsStatusBreakdown;
}

export type AdminListCampaignsParams = {
status?: AdminListCampaignsStatus;
sortBy?: AdminListCampaignsSortBy;
sortOrder?: AdminListCampaignsSortOrder;
page?: number;
limit?: number;
};

export type AdminListCampaignsStatus = typeof AdminListCampaignsStatus[keyof typeof AdminListCampaignsStatus];


export const AdminListCampaignsStatus = {
  processing: 'processing',
  complete: 'complete',
  failed: 'failed',
} as const;

export type AdminListCampaignsSortBy = typeof AdminListCampaignsSortBy[keyof typeof AdminListCampaignsSortBy];


export const AdminListCampaignsSortBy = {
  createdAt: 'createdAt',
  budget: 'budget',
  fitScore: 'fitScore',
} as const;

export type AdminListCampaignsSortOrder = typeof AdminListCampaignsSortOrder[keyof typeof AdminListCampaignsSortOrder];


export const AdminListCampaignsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;