# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Campaign Generator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    post:
      operationId: submitCampaign
      tags: [campaigns]
      summary: Submit intake form and generate campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignSubmitResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign result by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CampaignInput:
      type: object
      required:
        - fullName
        - email
        - companyName
        - productDescription
        - targetAudience
        - budget
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        companyName:
          type: string
          minLength: 1
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        productDescription:
          type: string
          minLength: 10
        targetAudience:
          type: string
          minLength: 5
        budget:
          type: number
          minimum: 1
    ChannelAllocation:
      type: object
      required:
        - channel
        - percentage
        - amount
      properties:
        channel:
          type: string
        percentage:
          type: number
        amount:
          type: number
    ABTestVariant:
      type: object
      required:
        - angle
        - type
        - hypothesis
        - primaryMetric
      properties:
        angle:
          type: string
        type:
          type: string
        hypothesis:
          type: string
        primaryMetric:
          type: string
    ABTestPlanItem:
      type: object
      required:
        - channel
        - variantA
        - variantB
      properties:
        channel:
          type: string
        variantA:
          $ref: "#/components/schemas/ABTestVariant"
        variantB:
          $ref: "#/components/schemas/ABTestVariant"
    CampaignIdea:
      type: object
      required:
        - title
        - description
        - rationale
      properties:
        title:
          type: string
        description:
          type: string
        rationale:
          type: string
    AdCopyItem:
      type: object
      required:
        - channel
        - headline
        - body
        - callToAction
      properties:
        channel:
          type: string
        headline:
          type: string
        body:
          type: string
        callToAction:
          type: string
    CampaignResult:
      type: object
      required:
        - channelMix
        - budgetAllocation
        - abTestPlan
        - fitScore
      properties:
        channelMix:
          type: array
          items:
            type: string
        budgetAllocation:
          type: array
          items:
            $ref: "#/components/schemas/ChannelAllocation"
        abTestPlan:
          type: array
          items:
            $ref: "#/components/schemas/ABTestPlanItem"
        campaignIdeas:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/CampaignIdea"
        adCopy:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/AdCopyItem"
        fitScore:
          type: integer
        riskLevel:
          type: ["string", "null"]
          description: "Deterministic risk assessment: low / medium / high"
        errorMessage:
          type: ["string", "null"]
    CampaignDetail:
      type: object
      required:
        - id
        - status
        - budget
        - createdAt
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [processing, complete, failed]
        budget:
          type: number
        productDescription:
          type: string
        targetAudience:
          type: string
        createdAt:
          type: string
        result:
          $ref: "#/components/schemas/CampaignResult"
    CampaignSubmitResponse:
      type: object
      required:
        - campaignId
        - status
      properties:
        campaignId:
          type: integer
        status:
          type: string
          enum: [processing, complete, failed]
        result:
          $ref: "#/components/schemas/CampaignResult"
...