# lib/api-zod/src/generated/types/aBTestPlanItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Campaign Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ABTestVariant } from './aBTestVariant';

export interface ABTestPlanItem {
  channel: string;
  variantA: ABTestVariant;
  variantB: ABTestVariant;
}