# lib/api-zod/src/generated/types/aBTestVariant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Campaign Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ABTestVariant {
  angle: string;
  type: string;
  hypothesis: string;
  primaryMetric: string;
}