# lib/api-zod/src/generated/types/campaignSubmitResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Campaign Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignResult } from './campaignResult';
import type { CampaignSubmitResponseStatus } from './campaignSubmitResponseStatus';

export interface CampaignSubmitResponse {
  campaignId: number;
  status: CampaignSubmitResponseStatus;
  result?: CampaignResult;
}