# lib/api-zod/src/generated/types/channelAllocation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Campaign Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface ChannelAllocation {
  channel: string;
  percentage: number;
  amount: number;
}